refactor: the form has been separated from the modal and is inserted via slot
