+++ app/components/app/modal.vue
@@ -1,0 +1,38 @@
+<script setup lang="ts" generic="T">
+  interface Props {
+    title?: string
+  }
+  defineProps<Props>()
+
+  const isOpen = defineModel<boolean>()
+</script>
+
+<template>
+  <UModal
+    v-model="isOpen"
+    width="25"
+  >
+    <UCard
+      :ui="{
+        header: {
+          padding: 'py-2',
+        },
+      }"
+    >
+      <template #header>
+        <div class="flex justify-between items-center">
+          <h1 class="text-2xl font-semibold">{{ title }}</h1>
+          <UButton
+            color="black"
+            icon="mdi-close"
+            size="xs"
+            variant="ghost"
+            @click="isOpen = false"
+          />
+        </div>
+      </template>
+      <slot />
+      <template #footer />
+    </UCard>
+  </UModal>
+</template>
